pyproject.toml
[project]
name = "phototracks"
version = "0.1.0"
description = "phototracks matches your photos with GPS locations based on timestamps from GPX track files"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name="Adrián Lattes", email="adrianlattes@disroot.org"},
]
dependencies = [
    "asyncio>=3.4.3",
    "click>=8.2.1",
    "exifread>=3.3.2",
    "fastapi>=0.116.1",
    "geopandas>=1.1.1",
    "gpx>=0.2.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Natural Language :: English",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
phototracks= "phototracks.__main__:main"

[dependency-groups]
dev = [
    "black",
    "commitizen",
    "flake8",
    "ipython>=8.37.0",
    "isort",
    "pre-commit",
    "pyright",
]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]
docs = [
    "markdown-include",
    "mkdocs",
    "mkdocs-autolinks-plugin",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-htmlproofer-plugin",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "mkdocs-section-index",
    "mkdocstrings",
    "mkdocstrings-python",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
addopts = "-ra --cov=src"

[tool.coverage.report]
exclude_also = [
    'pragma: no cover',
    'def __repr__',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    'class .*\bProtocol\):',
    '@(abc\.)?abstractmethod',
]
omit = [
    "*/.local/*",
    "/usr/*",
    "/nix/store/*",
    ]

[tool.commitizen]
version_provider = "uv"
tag_format =  "v${version}"
version_scheme = "semver"
update_changelog_on_bump = true
version_files = [
  "pyproject.toml:version",
  "src/phototracks/version.py",
]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.isort]
profile = "black"
src_paths = ["src", "test"]
